Refactor: Reduce module complexity and update dependencies
- Updated 10 dependencies to latest minor versions (security improvements)
- Reduced note.js complexity from 9 to 8 dependencies via modular extraction:
  * Created noteEvents.js for event handling (3 deps)
  * Created ghostConnectors.js for UI components (0 deps)
  * Created noteDeletion.js for deletion logic (4 deps)
- Reduced app.js complexity from 8 to 7 dependencies via separation:
  * Created uiSetup.js for UI initialization (5 deps)
  * Created canvasInitialization.js for canvas setup (4 deps)
- Improved architecture: 0 modules >8 dependencies (was 2)
- All unit tests passing (20/20 ✅)
- Health score: 86/100 maintained
- Better separation of concerns and maintainability
