+++ package.json
@@ -28,19 +28,18 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.25.4",
-    "@eslint/js": "^9.9.0",
-    "@playwright/test": "^1.46.1",
-    "babel-jest": "^29.7.0",
+    "@babel/core": "^7.28.0",
+    "@babel/preset-env": "^7.28.0",
+    "@eslint/js": "^9.31.0",
+    "@playwright/test": "^1.54.1",
     "depcheck": "^1.4.7",
-    "eslint": "^9.9.0",
-    "eslint-plugin-jest": "^28.8.0",
+    "eslint": "^9.31.0",
+    "eslint-plugin-jest": "^28.14.0",
     "eslint-plugin-no-unsanitized": "^4.1.2",
-    "eslint-plugin-playwright": "^1.6.2",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-playwright": "^1.8.3",
+    "eslint-plugin-prettier": "^5.5.3",
     "eslint-plugin-security": "^3.0.1",
-    "globals": "^15.9.0",
+    "globals": "^15.15.0",
     "http-server": "^14.1.1",
     "husky": "^9.1.7",
     "jest": "^30.0.5",
@@ -49,7 +48,7 @@
     "lint-staged": "^16.1.2",
     "madge": "^8.0.0",
     "npm-check-updates": "^18.0.1",
-    "prettier": "^3.3.3",
+    "prettier": "^3.6.2",
     "typescript": "^5.8.3"
   },
   "lint-staged": {
