+++ src/js/core/uiSetup.js
@@ -1,0 +1,111 @@
+// uiSetup.js - Handles UI initialization and setup
+import { log } from '../utils/utils.js';
+import { canvasManager } from './canvasManager.js';
+import { exportToJSON, importFromJSON } from '../data/dataStore.js';
+import { clearAllState } from '../data/storageManager.js';
+import { setupZoomAndPan, setFixedZoom } from '../features/zoom/zoomManager.js';
+
+export function setupUI(elements) {
+  populateCanvasStyleDropdown(elements);
+  setupExportImport(elements.menu, elements.canvas);
+  setupClearCanvas(elements.menu, elements.canvas);
+}
+
+function populateCanvasStyleDropdown(elements) {
+  canvasManager.getAvailableModules().forEach((moduleName) => {
+    const button = createDropdownButton(moduleName, () =>
+      switchCanvas(moduleName, elements),
+    );
+    elements.canvasStyleDropdown.appendChild(button);
+  });
+}
+
+function createDropdownButton(text, onClick) {
+  const li = document.createElement('li');
+  const button = document.createElement('button');
+  button.textContent = text;
+  button.addEventListener('click', onClick);
+  li.appendChild(button);
+  return li;
+}
+
+function switchCanvas(moduleName, elements) {
+  log('Switching to canvas:', moduleName);
+  try {
+    canvasManager.switchBackgroundLayout(moduleName, elements.canvas);
+
+    // Update canvas dimensions if necessary
+    const currentModule = canvasManager.getCurrentModule();
+    if (currentModule) {
+      elements.canvas.style.width = `${currentModule.width}px`;
+      elements.canvas.style.height = `${currentModule.height}px`;
+    }
+
+    // Handle zoom and pan
+    const centerX = elements.canvas.clientWidth / 2;
+    const centerY = elements.canvas.clientHeight / 2;
+    setFixedZoom(1, elements.canvas, elements.zoomDisplay, centerX, centerY);
+    setupZoomAndPan(
+      elements.canvasContainer,
+      elements.canvas,
+      elements.zoomDisplay,
+    );
+  } catch (error) {
+    console.error(`Error switching to canvas ${moduleName}:`, error);
+    alert(`Failed to switch to ${moduleName}. Please try again.`);
+  }
+}
+
+function setupExportImport(menu, canvas) {
+  const exportButton = menu.querySelector('#export-btn');
+  if (exportButton) {
+    exportButton.addEventListener('click', handleExportToFile);
+  }
+
+  const importButton = menu.querySelector('#import-btn');
+  if (importButton) {
+    importButton.addEventListener('change', () => handleImportFromFile(canvas));
+  }
+}
+
+function handleExportToFile() {
+  const json = exportToJSON();
+  const blob = new Blob([json], { type: 'application/json' });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = 'mindmap_export.json';
+  a.click();
+  URL.revokeObjectURL(url);
+}
+
+function handleImportFromFile(canvas) {
+  const input = document.createElement('input');
+  input.type = 'file';
+  input.accept = '.json';
+  input.onchange = (event) => {
+    const file = event.target.files[0];
+    const reader = new FileReader();
+    reader.onload = (e) => {
+      try {
+        importFromJSON(e.target.result, canvas);
+      } catch (error) {
+        console.error('Error importing file:', error);
+        alert("Error importing file. Please make sure it's a valid JSON file.");
+      }
+    };
+    reader.readAsText(file);
+  };
+  input.click();
+}
+
+function setupClearCanvas(menu, canvas) {
+  const clearButton = menu.querySelector('#clear-btn');
+  if (clearButton) {
+    clearButton.addEventListener('click', () => {
+      if (confirm('Are you sure you want to clear the canvas?')) {
+        clearAllState(canvas);
+      }
+    });
+  }
+}
