+++ src/js/features/note/ghostConnectors.js
@@ -1,0 +1,9 @@
+// ghostConnectors.js - Handles ghost connector creation for notes
+export function createGhostConnectors(note) {
+  const positions = ['top', 'bottom', 'left', 'right'];
+  positions.forEach((position) => {
+    const connector = document.createElement('div');
+    connector.className = `ghost-connector ${position}`;
+    note.appendChild(connector);
+  });
+}
