+++ src/js/features/note/noteDeletion.js
@@ -1,0 +1,20 @@
+// noteDeletion.js - Handles note deletion logic
+import { deleteNoteById } from '../../data/dataStore.js';
+import { deleteConnectionsByNote } from '../connection/connection.js';
+import { connectionManager } from '../connection/connectionManager.js';
+import { NoteManager } from '../../core/event.js';
+
+export function deleteNoteWithConnections(note, canvas) {
+  deleteConnectionsByNote(note);
+  deleteNoteById(note.id);
+  note.remove();
+  connectionManager.updateConnections(note, canvas);
+}
+
+export function deleteNote() {
+  const selected = NoteManager.getSelectedNotes()[0];
+  if (selected) {
+    const canvas = document.getElementById('canvas');
+    deleteNoteWithConnections(selected, canvas);
+  }
+}
