+++ src/js/utils/deviceUtils.js
@@ -1,0 +1,9 @@
+//src/js/utils/deviceUtils.js
+
+/**
+ * Detects if the current device is a mobile device.
+ * @returns {boolean} - True if the device is mobile, false otherwise.
+ */
+export function isMobileDevice() {
+  return /Mobi|Android/i.test(navigator.userAgent);
+}
