+++ src/js/utils/utils.js
@@ -105,11 +105,3 @@ export const fromBase62 = (str) => {
     .split('')
     .reduce((acc, char) => acc * 62 + BASE62.indexOf(char), 0);
 };
-
-/**
- * Detects if the current device is a mobile device.
- * @returns {boolean} - True if the device is mobile, false otherwise.
- */
-export function isMobileDevice() {
-  return /Mobi|Android/i.test(navigator.userAgent);
-}
